Avoid a cyclic import by moving some enums
